pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-agents-cli"
version = "0.1.0"
description = "CLI for Azure Agent Service"
authors = [ { name = "Your Name", email = "you@example.com" } ]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = ["azure", "agents", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.0",
  "azure-identity>=1.16.0",
  "azure-ai-projects>=1.0.0b6" # Azure AI Projects SDK for Agents
]

[project.scripts]
aza = "aza.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["aza*"]

[tool.typer]
auto_envvar_prefix = "AZA"
